Moved function to reproject.mosaicking
